feat: add support for Arch Linux packages
A minimal Arch Linux package is quite simple: it's a tarball containing
the installed contents of the package plus one extra file for metadata.

The only required metadata file is `.PKGINFO`, which contains core
metadata about the package: name, version, dependencies, etc.

Install hooks, if present, are stored in `.INSTALL`, which is a shell
script containing a function for each phase (pre-install, etc.); this is
unlike deb and RPM, which have one file per phase.

There is also usually `.MTREE`, which contains a summary of the files in
the package in mtree format. It is not required to install a package,
but it's needed for determining whether any files have been changed
after installation, which is pretty important.

The above files are created by this code.

Another common file is `.BUILDINFO`, containing some extra metadata
about the environment in which the package was built; that is not
created.

Also, packages are typically compressed with zstd (the current standard)
or xz (the previous one); there doesn't seem to be a high-quality,
pure-Go library for either of those, however, so we stick to the Go
standard library and use gzip, which still works fine.

This commit also introduces some helper code that should be useful for
other package formats, but we leave any potential refactors for the
future.