fix: Ensure globbed FileInfo always has correct size
When globbing a file that has a FileInfo set, we would reuse the pointer
to the original file's FileInfo even if the matched files' sizes are
different, causing deb to error when writing the data file due to
mismatching sizes.

Copy the FileInfo and recalculate its size when globbing. Add a test
case to check this scenario.

Fixes #316.