feat: Add "force_implicit_dirs" to dir/tree.
This allows users to specify a list of directories to treat akin
to the list of system directories, i.e. things that the constructed
package should *not* take ownership of in e.g. the rpm database.

Example use case:
1) user belongs to FooCorp
2) FooCorp has many packages to install under /opt/foocorp
   e.g. /opt/foocorp/service[A,B,C...]
3) Using tree w/ forced implicit dirs keeps generated package(s)
   from claiming ownership of /opt/foocorp so that the
   service[A,B,C...] packages can be bundled and installed
   separately without path conflicts.

The particular scenario I ran into this for was with RHEL8+ being
much pickier about each dir being owned by one and only one package.
(We pair this with having one "foocorp-dirs" package that just
 owns the shared dir creation, that's used as a pre-req for the
 other packages.)